requirements.txt
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.4
google-generativeai==0.3.2
python-dotenv==1.0.0
tabulate==0.9.0
streamlit==1.32.0
